decoupling of interdependencies through INavmapBus, creation of command and back functionality & refactoring
