+++ Explorer/Assets/DCL/Navmap/INavmapCommand.cs
@@ -1,0 +1,12 @@
+using Cysharp.Threading.Tasks;
+using System.Threading;
+
+namespace DCL.Navmap
+{
+    public interface INavmapCommand
+    {
+        UniTask ExecuteAsync(CancellationToken ct);
+
+        void Undo();
+    }
+}
