+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Analytics/DCL.Analytics.asmdef
@@ -27,7 +27,8 @@
         "GUID:543b8f091a5947a3880b7f2bca2358bd",
         "GUID:e11871a9242c44ec98ed3459ff7781d9",
         "GUID:8baf705856414dad9a73b3f382f1bc8b",
-        "GUID:254c118155004a75a699a9410d44311f"
+        "GUID:254c118155004a75a699a9410d44311f",
+        "GUID:006c0e0a70294dbba8a4cbcfb77e1f7d"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
